Refactor latest releases feed code.
The code has been refactored to directly re-use (using require_once to
directly include the code) as much formatting code from the main catalog
project as possible.

A bunch of work was recently done in the main catalog to update how
authors are displayed such as how multiple authors are displayed and better
formatting of the years. Instead of re-writing that logic here or copy and
pasting code from catalog it seemed better to directly "require" the code
here so that (some) future updates will propagate automatically (the
title_blade.php code is still re-implemented in the WordPress theme).

The drawback to this approach is that you could modify the catalog code
and not realize you just broke the WordPress theme. A comment should be
added to the relevant functions in the catalog code if this commit is
merged.